\name{Ytools}
\alias{Ytools}
\title{Tools for iron metabolism yeasts explorers}
\description{
  A suite of tools to explore multi-omics data in \emph{S. cerevisiae} and
  pathologic yeasts (\emph{C. glabrata} and \emph{C. albicans}).

  In this package, you can find :
  \describe{
    \item{\strong{annotations_C_albicans()}}{Get the last annotations of \emph{C. albicans} on CGD}
    \item{\strong{annotations_C_glabrata()}}{Get the last annotations of \emph{C. glabrata} on CGD}
    \item{\strong{annotationsS_cerevisiae()}}{Get the last annotations of \emph{S. cerevisiae} on SGD}
    \item{\strong{GO_CGD()}}{Get and read last GO term associated of CGD}
    \item{\strong{GO_S_cerevisiae()}}{Get and read last GO term associated of S. cerevisiae}
    \item{\strong{Selection_Annotations_L1()}}{Selection of a list of genes by annotation with a list of terms}
    \item{\strong{Selection_GO_L1()}}{Selection of a list of genes by GO description, with a list of terms}
    \item{\strong{Selection_GRYC_L1()}}{Selection of a list of genes by GO descriptio, with a list of terms}
  }
}
\author{
Maintainer: Thomas DENECKER

Author : Thomas DENECKER
}
